+++ src/reduce/index.ts
